New Version
This Version contain hundrets of fixes. Most important is the xplfatform QString handling with FoxSDK.
Else:
- New Ribbon handling
- Reorganized Icons

+++ src/nodes/qaudiotrackitem.cpp
@@ -1,6 +1,6 @@
 /*
  *  DiskButler - a powerful CD/DVD/BD recording software tool for Linux, macOS and Windows.
- *  Copyright (c) 20019 Ingo Foerster (pixbytesl@gmail.com).
+ *  Copyright (c) 2021 Ingo Foerster (pixbytesl@gmail.com).
  *
  *  This program is free software; you can redistribute it and/or modify
  *  it under the terms of the GNU General Public License 3 as published by
@@ -121,7 +121,7 @@ const int32* QAudioTrackItem::getPointerIndexes()
 
 int32 QAudioTrackItem::getIndexesLength()
 {
-  return mIndexes.size();
+  return static_cast<int>(mIndexes.size());
 }
 
 void QAudioTrackItem::resetAll()
